NV Lesson 4 homework v.2

+++ src/main/java/student_nikita_visnakovs/lesson_4/level_4/Calculator.java
@@ -16,7 +16,7 @@ public class Calculator {
         return firstNumber * secondNumber;
     }
 
-    public int div(int firstNumber, int secondNumber) {
+    public double div(double firstNumber, double secondNumber) {
         return firstNumber / secondNumber;
     }
